# openapi.yaml
openapi: 3.1.0
info:
  title: flags
  version: "1.0.0"
  description: >
    Feature flags and typed settings per environment: on/off state, percentage rollouts with
    stable per-user buckets, targeting rules on users/emails/attributes, change history, and
    cached ETag snapshots for client-side evaluation.


    Auth: every `/v1/*` and `/metrics` route requires `Authorization: Bearer <api-key>`
    (see `ApiKeyAuth` below). A key has a role (`read`, `write`, or `readwrite`) and an optional
    environment scope (`envs`: a list of environment names, or unscoped when the key applies to
    every environment). Any operation that names an `:env` path parameter, or an `env`/`to`
    body/query field, is environment-scope-gated: an environment-scoped key gets `403 FORBIDDEN`
    for any environment outside its scope, exactly like a role mismatch. This is called out again
    per operation below where it applies.


    Every response carries `Cache-Control: no-store` unless the operation documents its own
    `Cache-Control` header (only `GET /v1/snapshot/{env}` does).
servers:
  - url: /
security:
  - ApiKeyAuth: []
...
paths:
...
  /v1/environments:
    get:
      operationId: flags.environments.list
      summary: List environments visible to this key
      description: >
        Environments are configured via `FLAGS_ENVIRONMENTS`, intersected with the key's own
        environment scope.
      tags: [environments]
      x-stack-visibility: authenticated
      responses:
        "200":
          description: Visible environments with their current change-counter version.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [env, version]
                      properties:
                        env: { type: string }
                        version: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/flags:
    post:
      operationId: flags.flags.create
      summary: Create a flag
      description: >
        Creates the flag and seeds identical initial state in every configured environment
        (`enabled`, `value`, `offValue` default per `kind`, `percentage` 100, no rules).
        Requires a `write`-capable key.
      tags: [flags]
      x-stack-visibility: authenticated
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateFlagBody" }
      responses:
        "201":
          description: Flag created.
          headers:
            location: { schema: { type: string }, description: "`/v1/flags/{key}`" }
          content:
            application/json:
              schema:
                type: object
                required: [flag]
                properties:
                  flag: { $ref: "#/components/schemas/Flag" }
        "400": { $ref: "#/components/responses/ValidationOrInvalidValue" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/FlagExists" }
        "413": { $ref: "#/components/responses/ValueTooLarge" }
        "429": { $ref: "#/components/responses/RateLimited" }
    get:
      operationId: flags.flags.list
      summary: List flags
      description: Cursor-paginated. `cursor` must already be a valid flag key (schema-enforced), so a malformed cursor is rejected as `400 VALIDATION_FAILED`, not a domain error.
      tags: [flags]
      x-stack-visibility: authenticated
      parameters:
        - name: q
          in: query
          schema: { type: string, minLength: 1, maxLength: 120 }
          description: Substring filter on key/description.
        - name: tag
          in: query
          schema: { type: string, minLength: 1, maxLength: 40 }
        - name: kind
          in: query
          schema: { $ref: "#/components/schemas/FlagKind" }
        - name: archived
          in: query
          schema: { type: string, enum: ["true", "false"] }
        - name: limit
          in: query
          schema: { type: string, pattern: "^([1-9]|[1-9][0-9]|1[0-9][0-9]|200)$" }
          description: Default 50, max 200.
        - name: cursor
          in: query
          schema: { $ref: "#/components/schemas/FlagKey" }
      responses:
        "200":
          description: Page of flags.
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Flag" } }
                  nextCursor: { type: [string, "null"] }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/flags/{key}:
    parameters:
      - $ref: "#/components/parameters/KeyParam"
    get:
      operationId: flags.flags.get
      summary: Get a flag
      tags: [flags]
      x-stack-visibility: authenticated
      responses:
        "200":
          description: The flag.
          content:
            application/json:
              schema:
                type: object
                required: [flag]
                properties:
                  flag: { $ref: "#/components/schemas/Flag" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/FlagNotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
    patch:
      operationId: flags.flags.patch
      summary: Patch flag metadata
      description: >
        Updates description/tags/archived/reshuffle only — environment state is changed through
        `PATCH /v1/flags/{key}/envs/{env}`. `reshuffle: true` generates a new rollout salt (every
        percentage-rollout bucket reshuffles) and, like `archived` changing, bumps every
        environment's version counter. Requires a `write`-capable key.
      tags: [flags]
      x-stack-visibility: authenticated
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchFlagBody" }
      responses:
        "200":
          description: Updated flag.
          content:
            application/json:
              schema:
                type: object
                required: [flag]
                properties:
                  flag: { $ref: "#/components/schemas/Flag" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/FlagNotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
    delete:
      operationId: flags.flags.delete
      summary: Delete a flag
      description: >
        **Requires an unscoped API key** (`envs: null`) regardless of role: an environment-scoped
        key gets `403 FORBIDDEN` unconditionally, even if it has `write` access and the key's
        scope would otherwise cover every configured environment — the check is on `envs === null`,
        not on scope coverage. This is in addition to, and checked before, the normal `write`-role
        requirement.
      tags: [flags]
      x-stack-visibility: authenticated
      responses:
        "204":
          description: Deleted. No body.
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: >
            Either the key lacks `write` access, or the key is environment-scoped (see operation
            description) — both surface as `FORBIDDEN`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "404": { $ref: "#/components/responses/FlagNotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/flags/{key}/history:
    get:
      operationId: flags.flags.getHistory
      summary: History for one flag
      tags: [history]
      x-stack-visibility: authenticated
      parameters:
        - $ref: "#/components/parameters/KeyParam"
        - $ref: "#/components/parameters/HistoryLimit"
        - $ref: "#/components/parameters/HistoryBefore"
      responses:
        "200": { $ref: "#/components/responses/HistoryPage" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/FlagNotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/history:
    get:
      operationId: flags.history.list
      summary: History across every flag
      tags: [history]
      x-stack-visibility: authenticated
      parameters:
        - $ref: "#/components/parameters/HistoryLimit"
        - $ref: "#/components/parameters/HistoryBefore"
      responses:
        "200": { $ref: "#/components/responses/HistoryPage" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/flags/{key}/envs/{env}:
    parameters:
      - $ref: "#/components/parameters/KeyParam"
      - $ref: "#/components/parameters/EnvParam"
    get:
      operationId: flags.envs.get
      summary: Get one flag's state in one environment
      description: Environment-scope-gated on `{env}` — a key scoped away from this environment gets `403 FORBIDDEN`.
      tags: [flags]
      x-stack-visibility: authenticated
      responses:
        "200":
          description: Environment state.
          content:
            application/json:
              schema:
                type: object
                required: [env, state]
                properties:
                  env: { type: string }
                  state: { $ref: "#/components/schemas/EnvironmentState" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/FlagNotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
      # UNKNOWN_ENV is also reachable here, folded into the shared FlagNotFound/UnknownEnv 404 pair below.
    patch:
      operationId: flags.envs.patch
      summary: Patch one flag's state in one environment
      description: >
        Environment-scope-gated on `{env}`. Requires a `write`-capable key. `value`/`offValue` are
        validated against the flag's `kind`; `rules` are validated and normalized by the domain
        layer (ids deduplicated-checked, emails lower-cased, list sizes bounded) — this is why
        `INVALID_RULE` is reachable even though the wire-level JSON Schema for `rules` only checks
        "array, at most 100 items".
      tags: [flags]
      x-stack-visibility: authenticated
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchEnvBody" }
      responses:
        "200":
          description: Updated environment state.
          content:
            application/json:
              schema:
                type: object
                required: [env, state]
                properties:
                  env: { type: string }
                  state: { $ref: "#/components/schemas/EnvironmentState" }
        "400":
          description: One of `VALIDATION_FAILED` (malformed request shape), `INVALID_VALUE` (`value`/`offValue` does not match the flag's `kind`), or `INVALID_RULE` (a `rules` entry fails validation).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/FlagNotFound" }
        "413": { $ref: "#/components/responses/ValueTooLarge" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/flags/{key}/envs/{env}/copy:
    post:
      operationId: flags.envs.copy
      summary: Copy one environment's state onto another
      description: >
        Copies `enabled`, `value`, `offValue`, `percentage` and `rules` from `{env}` (the path,
        source) onto `to` (the body, destination) for the same flag. Both `{env}` and `to` are
        environment-scope-gated independently — a key must have access to both, or it gets
        `403 FORBIDDEN`. Requires a `write`-capable key.
      tags: [flags]
      x-stack-visibility: authenticated
      parameters:
        - $ref: "#/components/parameters/KeyParam"
        - $ref: "#/components/parameters/EnvParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [to]
              properties:
                to: { $ref: "#/components/schemas/EnvName", description: "Destination environment (env-scope-gated same as the path {env})." }
      responses:
        "200":
          description: Destination environment's new state.
          content:
            application/json:
              schema:
                type: object
                required: [env, state]
                properties:
                  env: { type: string, description: "Equal to `to` from the request body." }
                  state: { $ref: "#/components/schemas/EnvironmentState" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/FlagNotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/evaluate:
    post:
      operationId: flags.evaluate.post
      summary: Evaluate flags for a context (JSON body)
      description: >
        Environment-scope-gated on the body's `env`. Evaluates every non-archived flag in `env`
        unless `keys` narrows the set; an unknown requested key evaluates to
        `{ value: null, reason: "missing" }` (default shape) rather than an error. Response shape
        depends on `details`: see the two `EvaluateResult*` schemas below.
      tags: [evaluate]
      x-stack-visibility: authenticated
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EvaluateBody" }
      responses:
        "200": { $ref: "#/components/responses/EvaluateOk" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        # UNKNOWN_ENV folded into the FlagNotFound/UnknownEnv-style 404 pattern is N/A here:
        # env is validated before any flag lookup, so only UnknownEnv (404) applies, never FlagNotFound.
      x-unknown-env-response: "404 UNKNOWN_ENV — see components.responses.UnknownEnv"
    get:
      operationId: flags.evaluate.get
      summary: Evaluate flags for a context (query string)
      description: >
        Same behavior and response shapes as `POST /v1/evaluate`, expressed as query parameters.
        Environment-scope-gated on `env`.


        **Dynamic parameter pattern**: any query key matching `attrs.<name>` is folded into
        `context.attrs[<name>]` — OpenAPI cannot express an open-ended, dynamically-prefixed query
        parameter set natively, so this is documented here rather than as a fixed parameter list.
        `attrs.plan=pro` below is one illustrative example, not an exhaustive enumeration; any
        number of `attrs.<name>` pairs may be given. A query key that does not start with `attrs.`
        and is not one of `env`/`userId`/`email`/`keys`/`details` is accepted by the wire schema
        (open `additionalProperties`) but silently ignored by the handler.
      tags: [evaluate]
      x-stack-visibility: authenticated
      parameters:
        - name: env
          in: query
          required: true
          schema: { $ref: "#/components/schemas/EnvName" }
        - name: userId
          in: query
          schema: { type: string, minLength: 1, maxLength: 128 }
        - name: email
          in: query
          schema: { type: string, maxLength: 254 }
        - name: keys
          in: query
          schema: { type: string, maxLength: 4000 }
          description: Comma-separated flag keys to narrow the evaluated set.
        - name: details
          in: query
          schema: { type: string, enum: ["true", "false"] }
        - name: attrs.plan
          in: query
          required: false
          schema: { type: string, maxLength: 128 }
          description: >
            Example only. Any `attrs.<name>` query key is accepted and becomes
            `context.attrs[<name>]` — this is one instance of that open-ended pattern, not a fixed
            parameter.
      responses:
        "200": { $ref: "#/components/responses/EvaluateOk" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/snapshot/{env}:
    get:
      operationId: flags.snapshot.get
      summary: Full rule set of an environment, for client-side evaluation
      description: >
        Environment-scope-gated on `{env}`. Returns every non-archived flag's complete evaluable
        state, **including its rollout `salt`** — this is deliberate: the salt is what lets a
        trusted client reproduce the same stable percentage-rollout bucketing locally (see
        `Evaluator.bucket`), so it is a genuine, intentional part of this response, not leaked
        internal state.


        Strong `ETag` (`"<env>-<version>"`); send `If-None-Match` with a previously-seen ETag to
        get `304` with an empty body when nothing changed. This route also overrides the service's
        default `Cache-Control: no-store` with `private, max-age=0, must-revalidate` so a
        conditional-GET-capable client can revalidate instead of refetching unconditionally.
      tags: [snapshot]
      x-stack-visibility: authenticated
      parameters:
        - $ref: "#/components/parameters/EnvParam"
        - name: If-None-Match
          in: header
          required: false
          schema: { type: string }
          description: A previously-received `ETag` value; matching it returns `304`.
      responses:
        "200":
          description: Full snapshot.
          headers:
            etag: { schema: { type: string }, description: 'Strong ETag, `"<env>-<version>"`.' }
            cache-control: { schema: { type: string, const: "private, max-age=0, must-revalidate" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Snapshot" }
        "304":
          description: Not modified since the ETag given in `If-None-Match`. No body.
          headers:
            etag: { schema: { type: string } }
            cache-control: { schema: { type: string, const: "private, max-age=0, must-revalidate" } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/UnknownEnv" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/stats:
    get:
      operationId: flags.stats.get
      summary: Aggregate counts
      tags: [stats]
      x-stack-visibility: authenticated
      responses:
        "200":
          description: Flag and per-environment counters.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Stats" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: >
        `Authorization: Bearer <secret>`. Keys are configured as `id:secret[:role[:env+env]]` via
        `FLAGS_API_KEYS`. `role` is `read`, `write`, or `readwrite` (`readwrite` satisfies both a
        `read`- and a `write`-gated route). The optional trailing `env+env` list scopes the key to
        those environments only; omitting it (or a key with `envs: null`) means the key applies to
        every configured environment. A route that names an `:env` path parameter, or an
        `env`/`to` body/query field, checks this scope and returns `403 FORBIDDEN` for an
        out-of-scope environment. `DELETE /v1/flags/{key}` additionally requires an *unscoped* key
        regardless of role — see that operation's own description.
...
  responses:
    Unauthorized:
      description: Missing or invalid API key.
      headers:
        www-authenticate: { schema: { type: string, const: Bearer } }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example: { error: { code: UNAUTHORIZED, message: "missing or invalid API key" } }
    Forbidden:
      description: The key's role or environment scope does not permit this operation (`FORBIDDEN`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example: { error: { code: FORBIDDEN, message: "this API key has no read access" } }
...
  schemas:
...
    FlagKind:
      type: string
      enum: [boolean, string, number, json]
...